+++ cmd/main.go
@@ -1,16 +1,21 @@
 package main
 
 import (
-	"github.com/duosonic/go-strings-history/cmd/injector"
-	"github.com/duosonic/go-strings-history/internal/adaptor/infrastructure/config/db"
-	"github.com/duosonic/go-strings-history/internal/adaptor/infrastructure/router"
+	"github.com/duosonic62/go-strings-history/cmd/injector"
+	"github.com/duosonic62/go-strings-history/internal/adaptor/infrastructure/db"
+	"github.com/duosonic62/go-strings-history/internal/adaptor/infrastructure/db/migration"
+	"github.com/duosonic62/go-strings-history/internal/adaptor/infrastructure/router"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
+
 	r := router.Router
-	var webApp = injector.Initialize()
-	db.Migrate()
+	webApp := injector.Initialize()
+
+	// db settings
+	migration.Migrate()
+	db.InitDB()
 
 	// ユーザ登録API
 	r.POST("/user", func(context *gin.Context) { webApp.UserController.CreateUser(context) })
