+++ go.mod
@@ -1,22 +1,26 @@
-module github.com/duosonic/go-strings-history
+module github.com/duosonic62/go-strings-history
 
 go 1.13
 
 require (
-	github.com/friendsofgo/errors v0.9.2 // indirect
+	github.com/DATA-DOG/go-sqlmock v1.3.3 // indirect
+	github.com/friendsofgo/errors v0.9.2
 	github.com/gin-gonic/gin v1.5.0
 	github.com/go-sql-driver/mysql v1.4.1
-	github.com/google/logger v1.0.1
+	github.com/gofrs/uuid v3.2.0+incompatible // indirect
+	github.com/google/logger v1.0.1 // indirect
 	github.com/google/uuid v1.1.1
+	github.com/kat-co/vala v0.0.0-20170210184112-42e1d8b61f12
 	github.com/pelletier/go-toml v1.6.0 // indirect
 	github.com/rubenv/sql-migrate v0.0.0-20191213152630-06338513c237
 	github.com/spf13/afero v1.2.2 // indirect
 	github.com/spf13/cast v1.3.1 // indirect
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	github.com/spf13/viper v1.6.1 // indirect
+	github.com/spf13/viper v1.6.1
 	github.com/volatiletech/inflect v0.0.0-20170731032912-e7201282ae8d // indirect
-	github.com/volatiletech/sqlboiler v3.6.1+incompatible // indirect
+	github.com/volatiletech/null v8.0.0+incompatible
+	github.com/volatiletech/sqlboiler v3.6.1+incompatible
 	golang.org/x/sys v0.0.0-20191224085550-c709ea063b76 // indirect
 	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543 // indirect
 	gopkg.in/ini.v1 v1.51.1 // indirect
