+++ internal/domain/factory/user.go
@@ -1,6 +1,6 @@
 package factory
 
-import "github.com/duosonic/go-strings-history/internal/domain/entity"
+import "github.com/duosonic62/go-strings-history/internal/domain/entity"
 
 type UserFactory struct {
 	idFactory    IDFactory
