+++ internal/domain/repository/user.go
@@ -1,6 +1,6 @@
 package repository
 
-import "github.com/duosonic/go-strings-history/internal/domain/entity"
+import "github.com/duosonic62/go-strings-history/internal/domain/entity"
 
 type UserRepository interface {
 	Save(entity.User)
