+++ internal/usecase/interactor/user_command.go
@@ -1,13 +1,13 @@
 package interactor
 
 import (
-	"github.com/duosonic/go-strings-history/internal/domain/factory"
-	"github.com/duosonic/go-strings-history/internal/domain/repository"
-	"github.com/duosonic/go-strings-history/internal/usecase/inputboundary"
-	"github.com/duosonic/go-strings-history/internal/usecase/outputboundary"
-	"github.com/duosonic/go-strings-history/pkg/usecase/input"
-	"github.com/duosonic/go-strings-history/pkg/usecase/input/command"
-	"github.com/duosonic/go-strings-history/pkg/usecase/output"
+	"github.com/duosonic62/go-strings-history/internal/domain/factory"
+	"github.com/duosonic62/go-strings-history/internal/domain/repository"
+	"github.com/duosonic62/go-strings-history/internal/usecase/inputboundary"
+	"github.com/duosonic62/go-strings-history/internal/usecase/outputboundary"
+	"github.com/duosonic62/go-strings-history/pkg/usecase/input"
+	"github.com/duosonic62/go-strings-history/pkg/usecase/input/command"
+	"github.com/duosonic62/go-strings-history/pkg/usecase/output"
 )
 
 type UserUseCaseInteractor struct {
